消息模块，内容管理
报告列表api开始对接

+++ src/components/contentManage.vue
@@ -12,17 +12,21 @@
     </div>
 
     <div class="rightContent_" v-for="(item,index) in articlesAarry">
-      <span class="includeBtn" @click="releaseBtn(item.id)"><span>发布</span></span>
+      <span class="includeBtn" @click="releaseBtn(item.id,index)"><span>发布</span></span>
       <router-link :to="{ path: '/homePage/articleDetail', query: { id:item.id,edit:'1'}}">
         <div class="rightContent">
-          <p class="title_bar">
+          <!-- <p class="title_bar">
             <span>{{item.title}}<span>
-          </p>       
+          </p>  -->     
+          <p class="title_bar" style="padding-right: 160px;">
+            <span class="ellipsis" style="display:block">{{item.title}}<span>
+          </p>   
           <p class="title_bottom">
             <span>
-              <span class="bottom_item">来源：<span>{{item.source}}（<span style="color:#0000FF;display:inline" @click="goSomewhere">{{item.link}}</span>）</span></span>
+              <span class="bottom_item">来源：<span>{{item.source}}<span @click="goSomewhere" class="bottomLink ellipsis">（{{item.link}}）</span></span></span>
               <span class="bottom_item">类别：<span>{{item.type}}</span></span>
-              <span class="bottom_item">时间：<span>{{item.date}}</span></span>
+              <!-- <span class="bottom_item">时间：<span>{{item.isAdded==1?(item.time.toISOString().slice(0,10)):item.time}}</span></span> -->
+              <span class="bottom_item">时间：<span>{{item.time}}</span></span>
               <span class="bottom_item grey" @click="clickDel($event,item.id)" @mouseover="overDel" @mouseout="outDel"><img src="../../static/img/alert-delete.png" alt=""><span class="dele-tips">删除</span></span>
             </span>
           </p>
@@ -77,16 +81,19 @@ export default {
               'date':'2014-06-11',
           }
         ]
-      }
+      },
+      userSource:{},
+      userid:'',
     }
   },
   methods:{
     doThis:function(){
 
     },
-    releaseBtn(id){
+    releaseBtn(id,index){
       $.when(releaseArticle(id)).done(function(data){
         if(data.state=="0"){
+          window.location.reload();
         }
         else{
           alert(data.data);
@@ -115,18 +122,23 @@ export default {
       console.log(e);
       e.stopPropagation();
       e.preventDefault();
-      $.when(deleteArticle(id)).done(function(data){
-        if(data.state=="0"){
-        }
-        else{
-          alert(data.data);
-        }
-      })
+      if(confirm("确认删除该篇文章？")){
+        $.when(deleteArticle(id)).done(function(data){
+          if(data.state=="0"){
+            window.location.reload();
+          }
+          else{
+            alert(data.data);
+          }
+        })
+      }
     },
     handleInputClick:function(){
+      var that=this;
       that.pageNo=1;
       $.when(contentSearch(that.input2,that.pageNo)).done(function(data){
         if(data.state=="0"){
+          that.articlesAarry=[];
           that.insertData(data.data);
         }
         else{
@@ -144,7 +156,17 @@ export default {
       var res=data;
       var len=that.articlesAarry.length;
       if(res.list&&res.list.length!=0){
-        $(that.$refs.rightBottom).children('p').text('点击加载更多内容');
+        if(res.list.length<20){
+          $(that.$refs.rightBottom).children('p').text('暂无更多文章');
+          if(that.pageNo==1){
+            that.articlesAarry=[];
+          }
+          else{}
+        }
+        else{
+          $(that.$refs.rightBottom).children('p').text('点击加载更多内容');
+        }
+        // $(that.$refs.rightBottom).children('p').text('点击加载更多内容');
         res.list.map(function(value,index){
           that.articlesAarry.push(value);
         })
@@ -163,7 +185,7 @@ export default {
       this.pageNo=this.pageNo+1;
       var that=this;
       if(this.input2==""){
-        $.when(getContentList(that.pageNo)).done(function(data){
+        $.when(getContentList(this.userid,that.pageNo,'0')).done(function(data){
           if(data.state=="0"){
             that.insertData(data.data);
           }
@@ -190,9 +212,11 @@ export default {
       matchMenu();
     })
 
-    that.insertData(that.tableData);//localTest
-
-    $.when(getContentList(that.pageNo)).done(function(data){
+    // that.insertData(that.tableData);//localTest
+    this.userSource=JSON.parse(localStorage.getItem("userSource"));
+    // this.level=this.userSource?this.userSource.level:'';
+    this.userid=this.userSource?this.userSource.id:'';
+    $.when(getContentList(this.userid,that.pageNo,'0')).done(function(data){
       if(data.state=="0"){
         that.insertData(data.data);
       }
@@ -299,6 +323,7 @@ export default {
       width: 100%;
       // border-top: 2px dashed #eee;
       padding: 20px 0;
+      padding-bottom:25px;
       font-size:0;
       >span{
         color:#868686;
@@ -319,6 +344,17 @@ export default {
           display:none;
         }
         // width:33.3%;
+      }
+      .bottom_item{
+        position: relative;
+        .bottomLink{
+          color: rgb(0, 0, 255);
+          display: inline-block;
+          width: 80%;
+          position: absolute;
+          bottom: -20px;
+          left: 0;
+        }
       }
       .bottom_item:first-child{
         width:50%;
